requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==21.3.1
boto3==1.17.44
botocore==1.20.44
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
coverage==5.5
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-admin-interface==0.16.0
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-nose==1.4.7
django-storages==1.11.1
Faker==8.0.0
flake8==3.8.4
gunicorn==20.1.0
idna==3.1
iniconfig==1.1.1
isort==5.7.0
jmespath==0.10.0
mccabe==0.6.1
mixer==7.1.2
nose==1.3.7
packaging==20.9
pathspec==0.5.9
pep8==1.7.1
Pillow==8.1.2
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.1.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.2
whitenoise==5.2.0
zope.interface==5.3.0
